fix: Data Type

+++ INFEST_Project/Util/ExcelToJsonWizard.v1.0.6/loader_output/RageFangSkillTable.cs
@@ -29,31 +29,31 @@ public class RageFangSkillTable
     /// <summary>
     /// 1m이내일 경우
     /// </summary>
-    public float Distance_1P1M;
+    public int Distance_1P1M;
 
     /// <summary>
     /// 3m이내일 경우
     /// </summary>
-    public float Distance_1P3M;
+    public int Distance_1P3M;
 
     /// <summary>
     /// 5m이내일경우
     /// </summary>
-    public float Distance_1P5M;
+    public int Distance_1P5M;
 
     /// <summary>
     /// 1m이내일 경우
     /// </summary>
-    public float Distance_2P1M;
+    public int Distance_2P1M;
 
     /// <summary>
     /// 3m이내일 경우
     /// </summary>
-    public float Distance_2P3M;
+    public int Distance_2P3M;
 
     /// <summary>
     /// 5m이내일경우
     /// </summary>
-    public float Distance_2P5M;
+    public int Distance_2P5M;
 
 }
